+++ src/test/fixtures/custom/my-class.js
@@ -53,3 +53,6 @@ class MyClass {
    */
   renamed_collection_return_types() { }
 }
+
+class ExcludeThisClass {
+}